mandatory normed

+++ get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: smetzler <smetzler@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/25 09:42:31 by smetzler          #+#    #+#             */
-/*   Updated: 2021/09/15 16:13:40 by smetzler         ###   ########.fr       */
+/*   Updated: 2021/09/18 14:00:29 by smetzler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,15 +16,17 @@
 # include <unistd.h>
 # include <stdlib.h>
 # include <stdio.h>
+# include <stddef.h>
+
 
 char	*get_next_line(int fd);
-char	*ft_calloc(int size, char filler);
+char	*ft_calloc(ssize_t size, char filler);
 char	*ft_makereturn(char **tonext, int location);
-char	*ft_prepnext(char **tonext, int location, int size);
+char	*ft_prepnext(char **tonext, int location, ssize_t size);
 void	ft_free(char **line);
 int		ft_strlen(char *str);
 int		ft_strchr(char *tonext, char c, int flag);
-char	*ft_strnjoin(char *previous, char *line, int size);
+char	*ft_strnjoin(char *previous, char *line, ssize_t size);
 char	*ft_strndup(char *s1, int start, int length);
 
 #endif
